Move base64 and footer format validation to common module
